move ColorPreview to separate file
