+++ src/components/ColorPicker/index.ts
@@ -1,4 +1,5 @@
 export * from "./BrandColorPicker";
 export * from "./ColorInput";
 export * from "./ColorPicker";
+export * from "./ColorPreview";
 export * from "./CustomColorPicker";
